+++ crypto-ws-client/src/clients/zbg/zbg_swap.rs
@@ -102,8 +102,8 @@ impl<'a> OrderBookTopK for ZbgSwapWSClient<'a> {
     }
 }
 
-fn to_candlestick_raw_channel(pair: &str, interval: u32) -> String {
-    let valid_set: Vec<u32> = vec![
+fn to_candlestick_raw_channel(pair: &str, interval: usize) -> String {
+    let valid_set: Vec<usize> = vec![
         60, 180, 300, 900, 1800, 3600, 7200, 14400, 21600, 43200, 86400, 604800,
     ];
     if !valid_set.contains(&interval) {
